Refactor test_webapp into several files
This gets rid of the "image_type" notion introduced in 4e48d71 that
required test_all functions with a bunch of yields.

Instead, I'm taking advantage of the fact that nose will run tests only
in classes that start with "Test". This lets me add a whole new set of
tests simply by adding a new class that defines a different value for
IMAGE_CHANNELS or CROP_SIZE, etc.

This will be helpful when creating tests for #189.
